requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.1.2
pytesseract==0.3.4
requests==2.23.0
soupsieve==2.0
urllib3==1.25.9
Werkzeug==1.0.1
